refactor: Dao의 select 함수 책임 분리
